To perform prediction on Stock market is not easy. To perform prediction.
To perform prediction on Stock market is not easy. To perform
prediction.
Decision trees and Bayesian Networks are used. Decision Tree and
Bayesian Networks classifiers
are built using ID3 and Naïve Bayes algorithms respectively. Same
training/learning datasets of
five cases and a test dataset for prediction are provided to the
classifiers. Based on the predictions
made forecast accuracies were calculated using root mean square
deviation. Results of under and
over forecasting of the algorithms are also presented A comparison on
the performance of both
the algorithms were provided.